Add version control file for improved repository management
This commit introduces a version control file named VERSION to the repository, inspired by the versioning system used in the Linux Torvalds Linux repository Makefile.
(Learn form the best hacker)

The VERSION file contains version information for the Hyprland Dotfiles project, enhancing control and clarity over the repository versioning process. The version details are as follows:

- # SPDX-License-Identifier: MIT
- VERSION: 0
- PATCHLEVEL: 0
- SUBLEVEL: 2
- EXTRAVERSION:
- NAME: Hyprland Dotfiles

This configuration denotes version 0.0.2 of the project, without any additional extras. Utilizing this file facilitates better management of project versions compared to solely relying on Git tags without committing a specific version.

Including this version control file enables more effective tracking and management of future development and releases of the Hyprland Dotfiles project.

When it's time to bump the version before pushing the code, simply update the version information in the VERSION file, commit the changes, tag the current commit with the new version number, and push both the tag and the commit to the remote repository.