feat: add wlogout blur background and animation with custom CSS
I figured out that to create the blur background, I needed to take a screenshot of the screen with `grim`, then use `ImageMagick` to apply the blur effect to the screenshot. It’s a bit slow since the bash script takes a few seconds to run, but it works perfectly.

Once the blurred image was generated, I updated the wlogout CSS to use it as the background. Now the wlogout screen has a smooth, blurred effect.

I also added hover animations to the wlogout buttons to make more beautiful.

On top of that, I updated the Hyprland keybinding to launch wlogout and Waybar config.

Lastly, I updated the installation script to include the necessary package installations.