UmmItOS v0.6.0
chore(version): update versioning scheme and license information

Changed versioning format to MAJOR.MINOR.PATCH and updated the license from GPL-2.0 to GPL-3.0.

Updated the name to UmmItOS v0.6.0 and added a description for the first Hong Kong Linux distribution.

Alright, this marks the start of the first release of UmmItOS, which can function as both a distribution and a set of dotfiles.

Future plans include splitting the project into two usage paths:

1. Distro installation via ISO.
2. Dotfiles installation script.